+++ test/test-dist.mjs
@@ -1,5 +1,5 @@
 import * as tsLuxon from "../dist/ts-luxon.umd.js";
-import pkg from "../package.json" assert { type: "json" };
+import pkg from "../package.json" with { type: "json" };
 
 if (pkg.version !== tsLuxon.VERSION) {
     console.error(`Version mismatch! Expected ${pkg.version}, got ${tsLuxon.VERSION}. Rebuild the package to solve!`);